update paths and rename stubs to typings
